requirements.txt
asgiref==3.7.2
Django==4.2.4
django-browser-reload==1.11.0
django-fbv==0.5.1
PyYAML==6.0.1
slippers==0.6.2
sqlparse==0.4.4
typeguard==2.13.3
typing_extensions==4.7.1
tzdata==2023.3
